CI: setup QEMU to clobber unused vector elements for RVV
When mask-agnostic and tail-agnostic mode is set, QEMU's default
behaviour is to still act like the mode is set to
tail-undisturbed and mask-undisturbed, so a programmer could forget to
set the mode correctly when using masks or differing vector sizes,
and CI will still pass in some cases.

These QEMU_CPU flags toggle the behaviour so that *-agnostic modes
should be treated as if they always destroy data,
making it more likely to detect these problems, although QEMU might
not respect this fully for all operations at the time of writing.