standardize on a single `--target-arch` arg and use separate build dirs per target arch
This simplifies the various different naming conventions for targets and archs.
It also uses a separate build directory for each target arch,
as `cmake` and `meson` don't allow you to reconfigure it on the fly,
so we'd otherwise lose caching.

This also fixes a few hardcoded paths that were discovered as a result of this change.